Add new js files
