+++ src/index.js
@@ -1,52 +1,15 @@
-import Task from './buttons.js';
+import Task from './modules/buttons.js';
 import './style.css';
+import { renderTasks, addButton, addInput, clearAllButton } from './modules/todo.js';
 
-const todoList = document.querySelector('.todo-list');
-const addInput = document.querySelector('.add-input');
-const addButton = document.querySelector('.add-button');
-const clearAllButton = document.querySelector('.clear-all');
-
-function renderTasks() {
-  const tasks = Task.getTasks();
-  todoList.innerHTML = '';
-  tasks.forEach((task) => {
-    const li = document.createElement('li');
-    li.className = 'todo-item';
-    li.innerHTML = ` 
-      <label data-id=${task.id} class="${task.completed ? 'todo-completed' : ''}"> 
-      <input type="checkbox" class="todo-item-check" ${task.completed ? 'checked' : ''}> 
-      <input type="text" class="todo-item-edit" value="${task.description}">
-      </label>
-      <i class="icon"></i> 
-      `;
-    const editInput = li.querySelector('.todo-item-edit');
-    editInput.addEventListener('blur', () => {
-      const newDescription = editInput.value;
-      Task.editTaskDescription(task.id, newDescription);
-      renderTasks();
-    });
-
-    const checkbox = li.querySelector('.todo-item-check');
-    checkbox.addEventListener('change', () => {
-      Task.toggleTaskStatus(task.id);
-      renderTasks();
-    });
-
-    const deleteIcon = li.querySelector('.icon');
-    deleteIcon.addEventListener('click', () => {
-      Task.removeTask(task.id);
-      renderTasks();
-    });
-    todoList.appendChild(li);
-  });
-}
+renderTasks(Task);
 
 addButton.addEventListener('click', () => {
   const description = addInput.value;
   if (description) {
     Task.addTask(description);
     addInput.value = '';
-    renderTasks();
+    renderTasks(Task);
   }
 });
 
@@ -58,7 +21,5 @@ clearAllButton.addEventListener('click', () => {
   });
 
   Task.setTasks(newTasks);
-  renderTasks();
+  renderTasks(Task);
 });
-
-renderTasks();
